Move to_cvat_mask to a separate file
